--- api/gameRouter.js
@@ -1,15 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-const GET_ENDPOINT = '/games';
-const POST_ENDPOINT = '/games';
-
-router.get('/', (req, res) => {
-
-});
-
-router.post('/', (req, res) => {
-
-});
-
-module.exports = router;
