+++ api/server.spec.js
@@ -1,9 +1,31 @@
 const request = require('supertest');
 
-const server = require('./gameRouter');
+const server = require('./server');
 const data = require('../data');
 
-describe('game router endpoints', () => {
+const GAMES_GET_ENDPOINT = '/games';
+const GAMES_POST_ENDPOINT = '/games';
+
+describe('server endpoints', () => {
+
+    describe('get /', () => {
+
+        test('responds with 200', async () => {
+            const response = await request(server).get('/');
+            expect(response.status).toBe(200);
+        });
+
+        test('responds with json', async () => {
+            const response = await request(server).get('/');
+            expect(response.type).toMatch(/json/i);
+        });
+
+        test('sends correct response object', async () => {
+            const response = await request(server).get('/');
+            expect(response.body).toEqual({api: 'active'});
+        });
+
+    });
 
     describe('get /games', () => {
 
@@ -12,7 +34,7 @@ describe('game router endpoints', () => {
     describe('post /games', () => {
     
         afterEach(() => {
-            data = [];
+            data.clear();
         });
     
         test('responds with 201', async () => {
@@ -21,7 +43,7 @@ describe('game router endpoints', () => {
               genre: 'Arcade',
               releaseYear: 1980
             }
-            const response = await request(server).post(POST_ENDPOINT).send(body);
+            const response = await request(server).post(GAMES_POST_ENDPOINT).send(body);
             expect(response.status).toBe(201);
         });
     
@@ -31,7 +53,7 @@ describe('game router endpoints', () => {
               genre: 'Arcade',
               releaseYear: 1980
             }
-            const response = await request(server).post(POST_ENDPOINT).send(body);
+            const response = await request(server).post(GAMES_POST_ENDPOINT).send(body);
             expect(response.type).toMatch(/json/i);
         });
     
@@ -41,7 +63,7 @@ describe('game router endpoints', () => {
               genre: 'Arcade',
               releaseYear: 1980
             }
-            const response = await request(server).post(POST_ENDPOINT).send(body);
+            const response = await request(server).post(GAMES_POST_ENDPOINT).send(body);
             expect(response.body).toEqual({id: 1});
         });
     
@@ -50,7 +72,7 @@ describe('game router endpoints', () => {
                 genre: 'Arcade',
                 releaseYear: 1980
               }
-            const response = await request(server).post(POST_ENDPOINT).send(body);
+            const response = await request(server).post(GAMES_POST_ENDPOINT).send(body);
             expect(response.status).toBe(422);
         });
     
@@ -59,7 +81,7 @@ describe('game router endpoints', () => {
                 title: 'Pacman',
                 releaseYear: 1980
               }
-            const response = await request(server).post(POST_ENDPOINT).send(body);
+            const response = await request(server).post(GAMES_POST_ENDPOINT).send(body);
             expect(response.status).toBe(422);
         });
