+++ data.js
@@ -1,1 +1,15 @@
-module.exports = [];
+let games = []
+
+module.exports = {
+    add: (game) => {
+        return games.push(game);
+    },
+
+    clear: () => {
+        games = []
+    },
+
+    list: () => {
+        return games;
+    }
+}
