+++ index.js
@@ -1,19 +1,9 @@
 require('dotenv').config();
 const express = require('express');
 
-const gameRouter = require('./api/gameRouter');
-
-const server = express();
-server.use(express.json());
-server.use('/games', gameRouter);
-
-server.get('/', (req, res) => {
-    res.status(200).json({ api: 'active' });
-});
+const server = require('./api/server');
 
 const PORT = process.env.PORT || 5000;
 server.listen(PORT, () => {
     console.log(`\n** server up on port ${PORT} **\n`)
 });
-
-module.exports = server;
